+++ src/mailman/docs/NEWS.rst
@@ -103,7 +103,7 @@ Bugs
 * The ``gatenews`` command now parses messages with email.message_from_bytes
   without specifying policy.  (Closes #934)
 * Improve error message from REST API for invalid email address (Fixes #872)
-
+* Nonmember posts gated from usenet bypass nonmember moderation.  (Closes #937)
 
 Command line
 ------------